implementado preview de subida de documentos y solucion de errores

+++ solicitudes/views.py
@@ -8,6 +8,8 @@ from django.views.decorators.cache import never_cache
 from django.core.files.storage import FileSystemStorage
 from pathlib import Path
 from django.contrib import messages
+from django.http import HttpResponseForbidden
+
 
 from .forms import *
 from .models import *
@@ -16,6 +18,7 @@ from usuarios.views import verificarRedirect
 from modalidades.models import *
 from modalidades.forms import *
 
+from mensajes import notificaciones as notif
 # Create your views here.
 #########################################
 # Nota: Recordar
@@ -27,10 +30,13 @@ from modalidades.forms import *
 def verPDF(request, soli, file):
     documento_respuesta = get_object_or_404(RespuestaDocumento, pk=file)
     solicitud = get_object_or_404(Solicitud, pk=soli)
-
+    
     # Verificar si el documento pertenece a la solicitud
     if documento_respuesta.solicitud != solicitud:
         raise Http404("El recurso no existe")
+    if not((request.user.has_perm('permiso_administrador') and request.user.is_superuser == 1) or request.user.id == solicitud.solicitante_id ): 
+        return HttpResponseForbidden("No tienes los permisos necesarios para acceder a este documento.")
+    
 
     # Construir la ruta completa al archivo PDF
     ruta_path = Path(f'{settings.MEDIA_ROOT}{documento_respuesta.file}') 
@@ -67,6 +73,18 @@ def convocatorias(request):
     }
     return render(request, 'usuario_solicitud/convocatorias.html', context)
 
+def notificar_si_falta_documentos(solicitud):
+    documentos_modalidad = solicitud.modalidad.documento_set.all()
+    respuestas_documentos_solicitud = RespuestaDocumento.objects.filter(solicitud=solicitud) 
+    #si no hay el mismo numero de documentos que de respuesta enviar la notificacion
+    if set(documentos_modalidad) != set({resp.documento for resp in respuestas_documentos_solicitud}):        
+        notif.nueva(solicitud.solicitante, 'Falta uno o mas documentos requeridos en su solicitud. Favor de revisar su solicitud', 'solicitudes:documentos_convocatoria', urlArgs=[solicitud.modalidad_id])
+        solicitud.estado=Solicitud.ESTADO_CHOICES[1][0]
+        solicitud.save
+        return True
+    else:
+        return False
+
 def documentos_convocatorias(request, modalidad_id):
     solicitante = get_object_or_404(Usuario, pk=request.user.id)  
     url = verificarRedirect(solicitante)    
@@ -79,13 +97,16 @@ def documentos_convocatorias(request, modalidad_id):
         solicitante = usuario.solicitante
         modalidad = Modalidad.objects.get(pk = modalidad_id)
         documentos = Documento.objects.filter(modalidad__id=modalidad_id)
-        solicitud = Solicitud.objects.get(solicitante=solicitante, modalidad_id=modalidad, ciclo=ciclo_actual())
-        documentosResp = RespuestaDocumento.objects.filter(solicitud=solicitud)
-        listaDocumentos = zip(documentos, documentosResp) #Mete las dos listas de documentos y documentosRespuesta en una sola lista
+        solicitud = Solicitud.objects.get(solicitante=solicitante, modalidad_id=modalidad, ciclo=ciclo_actual())        
+        listaDocumentos = []
+        for documento in documentos:
+            tupla = (documento, RespuestaDocumento.objects.filter(solicitud=solicitud, documento=documento).first())
+            listaDocumentos.append(tupla)        
 
         convocatoria = Convocatoria.objects.all().first()
 
         context ={
+            'solicitud': solicitud,
             'modalidad': modalidad,
             'listaDocumentos': listaDocumentos,
             'convocatoria': convocatoria,
@@ -94,16 +115,42 @@ def documentos_convocatorias(request, modalidad_id):
         if request.method == 'POST':
             for doc in request.FILES:
                 documento = Documento.objects.get(id=doc) 
-                documentoRespuesta = RespuestaDocumento.objects.get(documento=doc, solicitud=solicitud)
-                files=request.FILES[doc] #Se obtiene el documento que se sube
-                print(files)
-                documentoRespuesta.file.delete() #Se elimina el documento que va a ser modificado
-                documentoRespuesta.file = files #Se agrega el nuevo documento
-                documentoRespuesta.save() 
+                try:
+                    documentoRespuesta = RespuestaDocumento.objects.get(documento=doc, solicitud=solicitud)
+                    files=request.FILES[doc] #Se obtiene el documento que se sube
+                    print(files)
+                    ext = os.path.splitext(files.name)[1][1:]                    
+                    if ext == 'pdf':
+                        documentoRespuesta.file.delete() #Se elimina el documento que va a ser modificado
+                        documentoRespuesta.file = files #Se agrega el nuevo documento
+                        documentoRespuesta.save() 
+                        messages.success(request, f"Exito al guardar el documento {documento.nombre}")
+                    else:
+                        messages.error(request, f'No se pudo guardar el documento {documento.nombre}: Sólo se permiten archivos en formato PDF.')                        
+                except RespuestaDocumento.DoesNotExist:
+                    print('Documento no existe')
+                    files={'file':request.FILES[doc]}
+                    documento = Documento.objects.get(id=doc)
+                    estado = RespuestaDocumento.ESTADO_CHOICES[0][0]
+                    data={
+                        'solicitud':solicitud,
+                        'documento': documento,
+                        'estado' : estado,
+                    }
+                    formRespDocs = DocumentoRespForm(data=data,files=files)
+                    
+                    if formRespDocs.is_valid():
+                        formRespDocs.save()
+                        messages.success(request, f"Exito al guardar el documento {documento.nombre}")
+                    else:
+                        messages.error(request, f'No se pudo guardar el documento {documento.nombre}')
+                        messages.error(request, formRespDocs.errors)
                 
-            messages.success(request, "Documentos modificados con éxito")
+                
+            #messages.success(request, "Documentos modificados con éxito")
+            notificar_si_falta_documentos(solicitud)
             return redirect("solicitudes:convocatorias")
-
+        
         return render(request, 'usuario_solicitud/modificar_docs_convocatoria.html', context)
       
     #Si el usuario aún no sube documentación para la modalidad elegida, se le muestra la vista para subir sus documentos
@@ -139,7 +186,7 @@ def documentos_convocatorias(request, modalidad_id):
             for doc in request.FILES:
                 files={'file':request.FILES[doc]}
                 documento = Documento.objects.get(id=doc)
-                estado = "pendiente"
+                estado = RespuestaDocumento.ESTADO_CHOICES[0][0]
                 data={
                     'solicitud':solicitud,
                     'documento': documento,
@@ -149,6 +196,11 @@ def documentos_convocatorias(request, modalidad_id):
                 
                 if formRespDocs.is_valid():
                     formRespDocs.save()
+                    messages.success(request, f"Exito al guardar el documento {documento.nombre}")
+                else:
+                    messages.error(request, f'No se pudo guardar el documento {documento.nombre}')
+                    messages.error(request, formRespDocs.errors)
+            notificar_si_falta_documentos(solicitud)
             messages.success(request, "Solicitud creada con éxito")
             return redirect("solicitudes:convocatorias")
